Change SensorConfig.interval from uint64 to uint32
The sampling interval was stored as a uint64, which is massively
overkill, as time is stored on the nodes as a uint32.

Change it to a uint32. This is a backwards compatible change.

+++ c/settings.pb.c
@@ -15,7 +15,7 @@ const pb_field_t SensorConfig_fields[8] = {
     PB_FIELD(  2, BOOL    , REQUIRED, STATIC  , OTHER, SensorConfig, hasADC2, hasADC1, 0),
     PB_FIELD(  3, BOOL    , REQUIRED, STATIC  , OTHER, SensorConfig, hasRain, hasADC2, 0),
     PB_FIELD(  4, UINT32  , OPTIONAL, STATIC  , OTHER, SensorConfig, avrID, hasRain, 0),
-    PB_FIELD(  5, UINT64  , REQUIRED, STATIC  , OTHER, SensorConfig, interval, avrID, 0),
+    PB_FIELD(  5, UINT32  , REQUIRED, STATIC  , OTHER, SensorConfig, interval, avrID, 0),
     PB_FIELD(  6, UENUM   , REQUIRED, STATIC  , OTHER, SensorConfig, routingMode, interval, 0),
     PB_FIELD(  7, UINT32  , OPTIONAL, STATIC  , OTHER, SensorConfig, powerID, routingMode, 0),
     PB_LAST_FIELD
